requirements.txt
aiohttp[speedups]
uvloop
ujson
aiogram
python-dotenv
